Moving structs about wto from metal to zos
